@@ -1,69 +1,40 @@
-import React, { Component } from 'react';
-import PropTypes from 'prop-types';
-import { View, Modal, Text } from 'react-native';
-import { Map } from 'immutable';
+import React, { useState, useEffect } from "react";
+import PropTypes from "prop-types";
+import { View, Modal, Text } from "react-native";
+import styles from "./styles";
 
-import styles from './styles';
+export default function Notification(props) {
+  const [message, setMessage] = useState(props.message);
 
-class Notification extends Component {
-  static propTypes = {
-    message: PropTypes.string,
-    duration: PropTypes.number.isRequired
+  useEffect(() => {
+    if (!message) {
+      setMessage(props.message);
+
+      const timer = setTimeout(() => {
+        setMessage(null);
+      }, props.duration);
+
+      return () => {
+        clearTimeout(timer);
+      };
+    }
+  }, [props.message]);
+
+  const modalProps = {
+    animationType: "fade",
+    transparent: true,
+    visible: Boolean(message)
   };
 
-  static defaultProps = {
-    duration: 1500
-  };
-
-  static getDerivedStateFromProps(props) {
-    // Update the "visible" state, based on whether
-    // or not there's a "message" value.
-    return {
-      ...this.state,
-      visible: Map([[null, false], [undefined, false]]).get(
-        props.message,
-        true
-      )
-    };
-  }
-
-  // The modal component is either "visible", or not.
-  // The "timer" is used to hide the notification
-  // after some predetermined amount of time.
-  state = { visible: false };
-  timer = null;
-
-  componentWillUnmount() {
-    clearTimeout(this.timer);
-  }
-
-  render() {
-    const modalProps = {
-      animationType: 'fade',
-      transparent: true,
-      visible: this.state.visible
-    };
-
-    this.timer = Map([
-      [null, () => null],
-      [undefined, () => null]
-    ]).get(this.props.message, () =>
-      setTimeout(
-        () => this.setState({ visible: false }),
-        this.props.duration
-      )
-    )();
-
-    return (
-      <Modal {...modalProps}>
-        <View style={styles.notificationContainer}>
-          <View style={styles.notificationInner}>
-            <Text>{this.props.message}</Text>
-          </View>
+  return (
+    <Modal {...modalProps}>
+      <View style={styles.notificationContainer}>
+        <View style={styles.notificationInner}>
+          <Text>{message}</Text>
         </View>
-      </Modal>
-    );
-  }
+      </View>
+    </Modal>
+  );
 }
 
 Notification.propTypes = {
@@ -74,5 +45,3 @@ Notification.propTypes = {
 Notification.defaultProps = {
   duration: 1500
 };
-
-export default Notification;
