@@ -1,19 +1,17 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import { View, Modal, ActivityIndicator } from 'react-native';
+import React from "react";
+import PropTypes from "prop-types";
+import { View, Modal, ActivityIndicator } from "react-native";
+import styles from "./styles";
 
-import styles from './styles';
-
-// The "Activity" component will only display
-// if the "visible" property is try. The modal
-// content is an "<ActivityIndicator>" component.
-const Activity = props => (
-  <Modal visible={props.visible} transparent>
-    <View style={styles.modalContainer}>
-      <ActivityIndicator size={props.size} />
-    </View>
-  </Modal>
-);
+export default function Activity(props) {
+  return (
+    <Modal visible={props.visible} transparent>
+      <View style={styles.modalContainer}>
+        <ActivityIndicator size={props.size} />
+      </View>
+    </Modal>
+  );
+}
 
 Activity.propTypes = {
   visible: PropTypes.bool.isRequired,
@@ -22,7 +20,5 @@ Activity.propTypes = {
 
 Activity.defaultProps = {
   visible: false,
-  size: 'large'
+  size: "large"
 };
-
-export default Activity;
