@@ -1,35 +1,29 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import { View, Text, Modal } from 'react-native';
+import React from "react";
+import PropTypes from "prop-types";
+import { View, Text, Modal } from "react-native";
+import styles from "./styles";
 
-import styles from './styles';
-
-// Declares styles for the error modal by
-// combining regular modal styles with
-// error styles.
 const innerViewStyle = [styles.modalInner, styles.modalInnerError];
-
 const textStyle = [styles.modalText, styles.modalTextError];
-
 const buttonStyle = [styles.modalButton, styles.modalButtonError];
 
-// Just like a success modal, accept for the addition of
-// error styles.
-const ErrorModal = props => (
-  <Modal {...props}>
-    <View style={styles.modalContainer}>
-      <View style={innerViewStyle}>
-        <Text style={textStyle}>Epic fail!</Text>
-        <Text style={buttonStyle} onPress={props.onPressConfirm}>
-          Fix it
-        </Text>
-        <Text style={buttonStyle} onPress={props.onPressCancel}>
-          Ignore it
-        </Text>
+export default function ErrorModal(props) {
+  return (
+    <Modal {...props}>
+      <View style={styles.modalContainer}>
+        <View style={innerViewStyle}>
+          <Text style={textStyle}>Epic fail!</Text>
+          <Text style={buttonStyle} onPress={props.onPressConfirm}>
+            Fix it
+          </Text>
+          <Text style={buttonStyle} onPress={props.onPressCancel}>
+            Ignore it
+          </Text>
+        </View>
       </View>
-    </View>
-  </Modal>
-);
+    </Modal>
+  );
+}
 
 ErrorModal.propTypes = {
   visible: PropTypes.bool.isRequired,
@@ -41,5 +35,3 @@ ErrorModal.defaultProps = {
   transparent: true,
   onRequestClose: () => {}
 };
-
-export default ErrorModal;
