@@ -1,25 +1,14 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import { ToastAndroid } from 'react-native';
-import { Map } from 'immutable';
+import React from "react";
+import PropTypes from "prop-types";
+import { ToastAndroid } from "react-native";
 
-// Toast helper. Always returns "null" so that the
-// output can be rendered as a React element.
-const show = (message, duration) => {
-  ToastAndroid.show(message, duration);
+export default function Notification({ message, duration }) {
+  if (message) {
+    ToastAndroid.show(message, duration);
+  }
+
   return null;
-};
-
-// This component will always return null,
-// since it's using an imperative React Native
-// interface to display popup text. If the
-// "message" property was provided, then
-// we display a message.
-const Notification = ({ message, duration }) =>
-  Map([[null, null], [undefined, null]]).get(
-    message,
-    show(message, duration)
-  );
+}
 
 Notification.propTypes = {
   message: PropTypes.string,
@@ -29,5 +18,3 @@ Notification.propTypes = {
 Notification.defaultProps = {
   duration: ToastAndroid.LONG
 };
-
-export default Notification;
