@@ -1,25 +1,25 @@
-import { StyleSheet } from 'react-native';
+import { StyleSheet } from "react-native";
 
 export default StyleSheet.create({
   container: {
     flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: 'ghostwhite',
+    justifyContent: "center",
+    alignItems: "center",
+    backgroundColor: "ghostwhite"
   },
 
   notificationContainer: {
     flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
+    justifyContent: "center",
+    alignItems: "center"
   },
 
   notificationInner: {
-    backgroundColor: 'azure',
+    backgroundColor: "azure",
     padding: 20,
     borderWidth: 1,
-    borderColor: 'lightsteelblue',
+    borderColor: "lightsteelblue",
     borderRadius: 2,
-    alignItems: 'center',
-  },
+    alignItems: "center"
+  }
 });
