@@ -1,13 +1,11 @@
-import React, { useEffect } from "react";
+import React from "react";
 import PropTypes from "prop-types";
 import { Alert } from "react-native";
 
 export default function ConfirmationAlert(props) {
-  useEffect(() => {
-    if (props.visible) {
-      Alert.alert(props.title, props.message, props.buttons);
-    }
-  });
+  if (props.visible) {
+    Alert.alert(props.title, props.message, props.buttons);
+  }
 
   return null;
 }
@@ -17,4 +15,10 @@ ConfirmationAlert.propTypes = {
   title: PropTypes.string,
   message: PropTypes.string,
   buttons: PropTypes.array
+};
+
+ConfirmationAlert.defaultProps = {
+  title: "",
+  message: "",
+  buttons: []
 };
